# Allow sampling from a closed integer range