--- a/src/distributions/mod.rs
+++ b/src/distributions/mod.rs
@@ -22,7 +22,7 @@ use Rng;
 pub use self::default::Default;
 pub use self::uniform::{uniform, codepoint, ascii_word_char};
 pub use self::uniform::{Uniform, Uniform01, Open01, Closed01, AsciiWordChar};
-pub use self::range::{Range};
+pub use self::range::Range;
 
 #[cfg(feature="std")]
 pub use self::gamma::{Gamma, ChiSquared, FisherF, StudentT};
@@ -39,7 +39,6 @@ mod uniform;
 mod ziggurat_tables;
 
 pub mod range;
-pub mod range2;
 
 #[cfg(feature="std")]
 pub mod gamma;