This does not compile, but I wanted to get your opinion.

The last few days I tried just about every trick to make sampling from a range of integers faster. Like making the zone a power of 2 (so mod becomes &), and multiplication with mersenne numbers. In the end, you are just trading the modulus for more times to call the prng (18~25% more). Because prng's can be much slower than the Xorshift used in the benchmarks, this is not a promising route to take...

It also took me a lot of time to understand [your method](https://github.com/dhardy/rand/commit/212e04e1e2eb7e06c8c70bf2c63acfa9945de03c) to calculate the zone. So I tried to come up with something myself, and it is exactly the same. I guess that proves there are no more off-by-one errors :-).

What I could do was add some extra comments to explain what is happening here. Also the modulus could be optimised a very little bit thanks to [this trick](https://stackoverflow.com/a/33333636): skipping it for the small chance `v` falls in the target range, making it a few percent faster.

What I would like to know your opinion about: is it useful to expose an inclusive range? As per [this comment](https://github.com/rust-lang/rfcs/pull/2106#issuecomment-324757225). It fit a little better with my mental model when writing the code.