@pitdicker 2017-08-30
I fixed the spelling in the comments, thank you.

About how or where to add a constructor: there are so many designs and questions in the RFC thread, I don't know where to start. And I now almost nothing about API design...
Probably best to leave some comments on the RFC tread first