@pitdicker 2017-09-01
I had two idea's that should help make the range code a bit faster in some cases:
- If the type of the range is u8 or u16, we should pick a zone that fits in an u32. This greatly reduces the number of random integers that have to be rejected. The zone can still be stored in the `RangeInt` struct, thanks to a trick: the bits of the u32 that get truncated, happen to all be 1's.
- There are two good techniques for this range code: the current one, that creates a zone as large as possible and reduces it with modulus. And one that reduces the random integers first to the power of two that is greater then the range, and then rejects everything greater than the range. What is optimal depends on how fast generating random numbers is compared to the modulus operation, and on how close a range is to its next power of two. We can just implement both techniques, and the pick the one that is probably best based on the size of the range. I think a good cutoff is when the range is more than 3/4 of the next power of two. Depending on the range, 0~25% of the numbers have to be rejected, 12.5% on average. But maybe 7/8 is better.

I'll report back when I have something.