@dhardy 2017-09-07
Yes, I had my head scratching why moving benchmarks from one module to another made a big difference, until I realised one used `blackbox`. Micro-benchmarks are tricky.

Can you add a `Range::new_inclusive` constructor?