@pitdicker 2017-09-28
Finally finished this. Sorry for taking so long.

Would it be okay if I make a PR that removes `Range` and replaces it with `Range2`?