@dhardy 2017-09-29
No problem.

Yes, I was planning on removing the original `range`; actually this PR is the reason I didn't yet. Go ahead.