@pitdicker 2017-10-14
Hi @dhardy. Is there a chance that you could merge this and the other PRs, or that we can move them along?