@dhardy 2017-10-14
Yeah, I guess. Sorry, I've been busy with some other work and travel the last couple of weeks, should have more time now.